*time.Location support Replaced deprecated packages and funcs
