+++ cleanenv_test.go
@@ -4,7 +4,7 @@ import (
 	"bytes"
 	"errors"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"net/url"
 	"os"
 	"reflect"
@@ -57,6 +57,7 @@ func TestReadEnvVars(t *testing.T) {
 		String          string            `env:"TEST_STRING"`
 		Duration        time.Duration     `env:"TEST_DURATION"`
 		Time            time.Time         `env:"TEST_TIME"`
+		Location        *time.Location    `env:"TEST_LOCATION"`
 		ArrayInt        []int             `env:"TEST_ARRAYINT"`
 		ArrayString     []string          `env:"TEST_ARRAYSTRING"`
 		MapStringInt    map[string]int    `env:"TEST_MAPSTRINGINT"`
@@ -75,7 +76,7 @@ func TestReadEnvVars(t *testing.T) {
 
 	type Required struct {
 		NotRequired int `env:"NOT_REQUIRED"`
-		Required    int `env:"REQUIRED" env-required:"true"`
+		Required    int `env:"REQUIRED"     env-required:"true"`
 	}
 
 	tests := []struct {
@@ -111,6 +112,7 @@ func TestReadEnvVars(t *testing.T) {
 				"TEST_STRING":          "test",
 				"TEST_DURATION":        "1h5m10s",
 				"TEST_TIME":            "2012-04-23T18:25:43.511Z",
+				"TEST_LOCATION":        "UTC",
 				"TEST_ARRAYINT":        "1,2,3",
 				"TEST_ARRAYSTRING":     "a,b,c",
 				"TEST_MAPSTRINGINT":    "a:1,b:2,c:3",
@@ -125,6 +127,7 @@ func TestReadEnvVars(t *testing.T) {
 				String:      "test",
 				Duration:    durationFunc("1h5m10s"),
 				Time:        timeFunc("2012-04-23T18:25:43.511Z", time.RFC3339),
+				Location:    time.UTC,
 				ArrayInt:    []int{1, 2, 3},
 				ArrayString: []string{"a", "b", "c"},
 				MapStringInt: map[string]int{
@@ -565,12 +568,12 @@ func TestParseFile(t *testing.T) {
 		Two int `yaml:"two" json:"two" toml:"two"`
 	}
 	type config struct {
-		Number  int64        `yaml:"number" json:"number" toml:"number"`
-		Float   float64      `yaml:"float" json:"float" toml:"float"`
-		String  string       `yaml:"string" json:"string" toml:"string"`
+		Number  int64        `yaml:"number"  json:"number"  toml:"number"`
+		Float   float64      `yaml:"float"   json:"float"   toml:"float"`
+		String  string       `yaml:"string"  json:"string"  toml:"string"`
 		Boolean bool         `yaml:"boolean" json:"boolean" toml:"boolean"`
-		Object  configObject `yaml:"object" json:"object" toml:"object"`
-		Array   []int        `yaml:"array" json:"array" toml:"array"`
+		Object  configObject `yaml:"object"  json:"object"  toml:"object"`
+		Array   []int        `yaml:"array"   json:"array"   toml:"array"`
 	}
 
 	wantConfig := config{
@@ -660,7 +663,7 @@ two = 2`,
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			tmpFile, err := ioutil.TempFile(os.TempDir(), fmt.Sprintf("*.%s", tt.ext))
+			tmpFile, err := os.CreateTemp(os.TempDir(), fmt.Sprintf("*.%s", tt.ext))
 			if err != nil {
 				t.Fatal("cannot create temporary file:", err)
 			}
@@ -730,7 +733,7 @@ func TestParseFileEnv(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			tmpFile, err := ioutil.TempFile(os.TempDir(), "*.env")
+			tmpFile, err := os.CreateTemp(os.TempDir(), "*.env")
 			if err != nil {
 				t.Fatal("cannot create temporary file:", err)
 			}
@@ -767,8 +770,8 @@ func TestParseFileEnv(t *testing.T) {
 
 func TestGetDescription(t *testing.T) {
 	type testSingleEnv struct {
-		One   int `env:"ONE" env-description:"one"`
-		Two   int `env:"TWO" env-description:"two"`
+		One   int `env:"ONE"   env-description:"one"`
+		Two   int `env:"TWO"   env-description:"two"`
 		Three int `env:"THREE" env-description:"three"`
 	}
 
@@ -778,8 +781,8 @@ func TestGetDescription(t *testing.T) {
 	}
 
 	type testDefaultEnv struct {
-		One   int `env:"ONE" env-description:"one" env-default:"1"`
-		Two   int `env:"TWO" env-description:"two" env-default:"2"`
+		One   int `env:"ONE"   env-description:"one"   env-default:"1"`
+		Two   int `env:"TWO"   env-description:"two"   env-default:"2"`
 		Three int `env:"THREE" env-description:"three" env-default:"3"`
 	}
 
@@ -899,8 +902,8 @@ func TestGetDescription(t *testing.T) {
 
 func TestFUsage(t *testing.T) {
 	type testSingleEnv struct {
-		One   int `env:"ONE" env-description:"one"`
-		Two   int `env:"TWO" env-description:"two"`
+		One   int `env:"ONE"   env-description:"one"`
+		Two   int `env:"TWO"   env-description:"two"`
 		Three int `env:"THREE" env-description:"three"`
 	}
 
@@ -975,7 +978,7 @@ func TestFUsage(t *testing.T) {
 			}
 			var cfg testSingleEnv
 			FUsage(w, &cfg, tt.headerText, uFuncs...)()
-			gotRaw, _ := ioutil.ReadAll(w)
+			gotRaw, _ := io.ReadAll(w)
 			got := string(gotRaw)
 
 			if got != tt.want {
@@ -987,10 +990,10 @@ func TestFUsage(t *testing.T) {
 
 func TestReadConfig(t *testing.T) {
 	type config struct {
-		Number    int64  `edn:"number" yaml:"number" env:"TEST_NUMBER" env-default:"1"`
-		String    string `edn:"string" yaml:"string" env:"TEST_STRING" env-default:"default"`
+		Number    int64  `edn:"number"     yaml:"number"     env:"TEST_NUMBER"     env-default:"1"`
+		String    string `edn:"string"     yaml:"string"     env:"TEST_STRING"     env-default:"default"`
 		NoDefault string `edn:"no-default" yaml:"no-default" env:"TEST_NO_DEFAULT"`
-		NoEnv     string `edn:"no-env" yaml:"no-env" env-default:"default"`
+		NoEnv     string `edn:"no-env"     yaml:"no-env"     env-default:"default"`
 	}
 
 	tests := []struct {
@@ -1137,7 +1140,7 @@ no-env: this
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			tmpFile, err := ioutil.TempFile(os.TempDir(), fmt.Sprintf("*.%s", tt.ext))
+			tmpFile, err := os.CreateTemp(os.TempDir(), fmt.Sprintf("*.%s", tt.ext))
 			if err != nil {
 				t.Fatal("cannot create temporary file:", err)
 			}
@@ -1161,5 +1164,26 @@ no-env: this
 				t.Errorf("wrong data %v, want %v", &cfg, tt.want)
 			}
 		})
+	}
+}
+
+// *time.Location is pointer type, so we need to compare it with pointer
+// reflect.DeepEqual() compares only pointer values, not structs
+func TestTimeLocation(t *testing.T) {
+	want := time.UTC
+
+	var S struct {
+		Location *time.Location `env:"TEST_LOCATION"`
+	}
+
+	os.Setenv("TEST_LOCATION", "UTC")
+	defer os.Clearenv()
+
+	if err := ReadEnv(&S); err != nil {
+		t.Fatal("cannot read env:", err)
+	}
+
+	if want != S.Location {
+		t.Errorf("wrong location %v, want %v", S.Location, want)
 	}
 }
